refactor(completion): split completion.rs and backend.rs into submodules
- Convert completion.rs (3156 lines) into completion/ module directory:
  - keyword.rs: PHP_KEYWORDS, keyword_completions, magic constants
  - symbols.rs: symbol_completions, builtin_completions, superglobals
  - member.rs: all_instance_members, all_static_members, resolve helpers
  - namespace.rs: collect_classes_with_ns, use_insert_position, FQN helpers
  - mod.rs: CompletionCtx, filtered_completions_at, shared helpers, tests
- Extract use-import helpers from backend.rs into use_import.rs:
  find_fqn_for_class, build_use_import_edit, find_use_insert_line

No behavior changes; all 686 tests pass.

Closes #97
