+++ src/completion/mod.rs
@@ -1,20 +1,37 @@
+mod keyword;
+pub use keyword::{keyword_completions, magic_constant_completions};
+
+mod symbols;
+pub use symbols::{
+    builtin_completions, superglobal_completions, symbol_completions, symbol_completions_before,
+};
+
+mod member;
+use member::{
+    all_instance_members, all_static_members, magic_method_completions, resolve_receiver_class,
+    resolve_static_receiver,
+};
+
+mod namespace;
+use namespace::{
+    collect_classes_with_ns, collect_fqns_with_prefix, current_file_namespace, typed_prefix,
+    use_completion_prefix, use_insert_position,
+};
+
 use std::sync::Arc;
 
-use php_ast::{ClassMemberKind, EnumMemberKind, ExprKind, NamespaceBody, Stmt, StmtKind};
 use tower_lsp::lsp_types::{
     CompletionItem, CompletionItemKind, InsertTextFormat, Position, Range, TextEdit, Url,
 };
 
 use tower_lsp::lsp_types::{Documentation, MarkupContent, MarkupKind};
 
-use crate::ast::{ParsedDoc, format_type_hint, offset_to_position};
+use crate::ast::{ParsedDoc, format_type_hint};
 use crate::docblock::find_docblock;
 use crate::hover::format_params_str;
 use crate::phpstorm_meta::PhpStormMeta;
-use crate::stubs::builtin_class_members;
 use crate::type_map::{
-    TypeMap, enclosing_class_at, is_backed_enum, is_enum, members_of_class, mixin_classes_of,
-    params_of_function, params_of_method, parent_class_name,
+    TypeMap, enclosing_class_at, members_of_class, params_of_function, params_of_method,
 };
 use crate::util::{camel_sort_key, fuzzy_camel_match, utf16_offset_to_byte};
 use std::collections::HashMap;
@@ -112,833 +129,6 @@ fn docblock_docs(doc: &ParsedDoc, sym_name: &str) -> Option<Documentation> {
     }
 }
 
-const PHP_KEYWORDS: &[&str] = &[
-    "abstract",
-    "and",
-    "array",
-    "as",
-    "break",
-    "callable",
-    "case",
-    "catch",
-    "class",
-    "clone",
-    "const",
-    "continue",
-    "declare",
-    "default",
-    "die",
-    "do",
-    "echo",
-    "else",
-    "elseif",
-    "empty",
-    "enddeclare",
-    "endfor",
-    "endforeach",
-    "endif",
-    "endswitch",
-    "endwhile",
-    "enum",
-    "eval",
-    "exit",
-    "extends",
-    "final",
-    "finally",
-    "fn",
-    "for",
-    "foreach",
-    "function",
-    "global",
-    "goto",
-    "if",
-    "implements",
-    "include",
-    "include_once",
-    "instanceof",
-    "insteadof",
-    "interface",
-    "isset",
-    "list",
-    "match",
-    "namespace",
-    "new",
-    "null",
-    "or",
-    "print",
-    "private",
-    "protected",
-    "public",
-    "readonly",
-    "require",
-    "require_once",
-    "return",
-    "self",
-    "static",
-    "switch",
-    "throw",
-    "trait",
-    "true",
-    "false",
-    "try",
-    "use",
-    "var",
-    "while",
-    "xor",
-    "yield",
-];
-
-pub fn keyword_completions() -> Vec<CompletionItem> {
-    PHP_KEYWORDS
-        .iter()
-        .map(|kw| CompletionItem {
-            label: kw.to_string(),
-            kind: Some(CompletionItemKind::KEYWORD),
-            ..Default::default()
-        })
-        .collect()
-}
-
-const PHP_MAGIC_CONSTANTS: &[(&str, &str)] = &[
-    ("__FILE__", "Absolute path of the current file"),
-    ("__DIR__", "Directory of the current file"),
-    ("__LINE__", "Current line number"),
-    ("__CLASS__", "Current class name"),
-    ("__FUNCTION__", "Current function name"),
-    ("__METHOD__", "Current method name (Class::method)"),
-    ("__NAMESPACE__", "Current namespace"),
-    ("__TRAIT__", "Current trait name"),
-];
-
-pub fn magic_constant_completions() -> Vec<CompletionItem> {
-    PHP_MAGIC_CONSTANTS
-        .iter()
-        .map(|(name, doc)| CompletionItem {
-            label: name.to_string(),
-            kind: Some(CompletionItemKind::CONSTANT),
-            detail: Some(doc.to_string()),
-            ..Default::default()
-        })
-        .collect()
-}
-
-pub fn symbol_completions(doc: &ParsedDoc) -> Vec<CompletionItem> {
-    let mut items = Vec::new();
-    collect_from_statements_with_doc(&doc.program().stmts, &mut items, Some(doc));
-    items
-}
-
-/// Like `symbol_completions` but only includes variables declared at or before `line`.
-/// Non-variable items (functions, classes, etc.) are always included.
-pub fn symbol_completions_before(doc: &ParsedDoc, line: u32) -> Vec<CompletionItem> {
-    let mut items = Vec::new();
-    collect_from_statements_before(
-        &doc.program().stmts,
-        &mut items,
-        line,
-        doc.source(),
-        Some(doc),
-    );
-    items
-}
-
-fn collect_from_statements_before(
-    stmts: &[Stmt<'_, '_>],
-    items: &mut Vec<CompletionItem>,
-    line: u32,
-    source: &str,
-    doc: Option<&ParsedDoc>,
-) {
-    for stmt in stmts {
-        match &stmt.kind {
-            StmtKind::Expression(e) => {
-                // Only add variables if they appear at or before the cursor line
-                let stmt_line = offset_to_position(source, stmt.span.start).line;
-                if stmt_line <= line {
-                    collect_from_expression(e, items);
-                }
-            }
-            StmtKind::Namespace(ns) => {
-                if let NamespaceBody::Braced(inner) = &ns.body {
-                    collect_from_statements_before(inner, items, line, source, doc);
-                }
-            }
-            // Non-variable items: always include
-            _ => {
-                collect_from_statements_with_doc(std::slice::from_ref(stmt), items, doc);
-            }
-        }
-    }
-}
-
-fn collect_from_statements_with_doc(
-    stmts: &[Stmt<'_, '_>],
-    items: &mut Vec<CompletionItem>,
-    doc: Option<&ParsedDoc>,
-) {
-    for stmt in stmts {
-        match &stmt.kind {
-            StmtKind::Function(f) => {
-                let sig = build_function_sig(f.name, &f.params, f.return_type.as_ref());
-                let documentation = doc.and_then(|d| docblock_docs(d, f.name));
-                let mut item =
-                    callable_item(f.name, CompletionItemKind::FUNCTION, !f.params.is_empty());
-                item.detail = Some(sig);
-                item.documentation = documentation;
-                items.push(item);
-                if let Some(named) = named_arg_item(f.name, CompletionItemKind::FUNCTION, &f.params)
-                {
-                    items.push(named);
-                }
-                for param in f.params.iter() {
-                    items.push(CompletionItem {
-                        label: format!("${}", param.name),
-                        kind: Some(CompletionItemKind::VARIABLE),
-                        ..Default::default()
-                    });
-                }
-            }
-            StmtKind::Class(c) => {
-                let class_name = c.name.unwrap_or("");
-                if !class_name.is_empty() {
-                    items.push(CompletionItem {
-                        label: class_name.to_string(),
-                        kind: Some(CompletionItemKind::CLASS),
-                        ..Default::default()
-                    });
-                }
-                for member in c.members.iter() {
-                    match &member.kind {
-                        ClassMemberKind::Method(m) => {
-                            let sig = build_function_sig(m.name, &m.params, m.return_type.as_ref());
-                            let documentation = doc.and_then(|d| docblock_docs(d, m.name));
-                            let mut item = callable_item(
-                                m.name,
-                                CompletionItemKind::METHOD,
-                                !m.params.is_empty(),
-                            );
-                            item.detail = Some(sig);
-                            item.documentation = documentation;
-                            items.push(item);
-                            if let Some(named) =
-                                named_arg_item(m.name, CompletionItemKind::METHOD, &m.params)
-                            {
-                                items.push(named);
-                            }
-                        }
-                        ClassMemberKind::Property(p) => {
-                            items.push(CompletionItem {
-                                label: format!("${}", p.name),
-                                kind: Some(CompletionItemKind::PROPERTY),
-                                ..Default::default()
-                            });
-                        }
-                        ClassMemberKind::ClassConst(c) => {
-                            items.push(CompletionItem {
-                                label: c.name.to_string(),
-                                kind: Some(CompletionItemKind::CONSTANT),
-                                ..Default::default()
-                            });
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            StmtKind::Interface(i) => {
-                items.push(CompletionItem {
-                    label: i.name.to_string(),
-                    kind: Some(CompletionItemKind::INTERFACE),
-                    ..Default::default()
-                });
-            }
-            StmtKind::Trait(t) => {
-                items.push(CompletionItem {
-                    label: t.name.to_string(),
-                    kind: Some(CompletionItemKind::CLASS),
-                    ..Default::default()
-                });
-            }
-            StmtKind::Enum(e) => {
-                items.push(CompletionItem {
-                    label: e.name.to_string(),
-                    kind: Some(CompletionItemKind::ENUM),
-                    ..Default::default()
-                });
-                for member in e.members.iter() {
-                    if let EnumMemberKind::Case(c) = &member.kind {
-                        items.push(CompletionItem {
-                            label: format!("{}::{}", e.name, c.name),
-                            kind: Some(CompletionItemKind::ENUM_MEMBER),
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            StmtKind::Namespace(ns) => {
-                if let NamespaceBody::Braced(inner) = &ns.body {
-                    collect_from_statements_with_doc(inner, items, doc);
-                }
-            }
-            StmtKind::Expression(e) => {
-                collect_from_expression(e, items);
-            }
-            _ => {}
-        }
-    }
-}
-
-fn collect_from_expression(expr: &php_ast::Expr<'_, '_>, items: &mut Vec<CompletionItem>) {
-    if let ExprKind::Assign(assign) = &expr.kind {
-        match &assign.target.kind {
-            ExprKind::Variable(name) => {
-                let label = format!("${}", name);
-                if label != "$this" {
-                    items.push(CompletionItem {
-                        label,
-                        kind: Some(CompletionItemKind::VARIABLE),
-                        ..Default::default()
-                    });
-                }
-            }
-            // Array destructuring: [$a, $b] = ... or list($a, $b) = ...
-            ExprKind::Array(elements) => {
-                for elem in elements.iter() {
-                    if let ExprKind::Variable(name) = &elem.value.kind {
-                        let label = format!("${}", name);
-                        if label != "$this" {
-                            items.push(CompletionItem {
-                                label,
-                                kind: Some(CompletionItemKind::VARIABLE),
-                                ..Default::default()
-                            });
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-        collect_from_expression(assign.value, items);
-    }
-}
-
-const PHP_BUILTINS: &[&str] = &[
-    // string
-    "strlen",
-    "strpos",
-    "strrpos",
-    "substr",
-    "str_replace",
-    "str_contains",
-    "str_starts_with",
-    "str_ends_with",
-    "str_split",
-    "explode",
-    "implode",
-    "join",
-    "trim",
-    "ltrim",
-    "rtrim",
-    "strtolower",
-    "strtoupper",
-    "ucfirst",
-    "lcfirst",
-    "ucwords",
-    "sprintf",
-    "printf",
-    "vsprintf",
-    "number_format",
-    "nl2br",
-    "htmlspecialchars",
-    "htmlentities",
-    "strip_tags",
-    "addslashes",
-    "stripslashes",
-    "str_pad",
-    "str_repeat",
-    "str_word_count",
-    "strcmp",
-    "strcasecmp",
-    "strncmp",
-    "strncasecmp",
-    "substr_count",
-    "substr_replace",
-    "strstr",
-    "stristr",
-    "preg_match",
-    "preg_match_all",
-    "preg_replace",
-    "preg_split",
-    "preg_quote",
-    "md5",
-    "sha1",
-    "hash",
-    "base64_encode",
-    "base64_decode",
-    "urlencode",
-    "urldecode",
-    "rawurlencode",
-    "rawurldecode",
-    "http_build_query",
-    "parse_str",
-    "parse_url",
-    // array
-    "count",
-    "array_key_exists",
-    "in_array",
-    "array_search",
-    "array_merge",
-    "array_replace",
-    "array_push",
-    "array_pop",
-    "array_shift",
-    "array_unshift",
-    "array_splice",
-    "array_slice",
-    "array_chunk",
-    "array_combine",
-    "array_diff",
-    "array_intersect",
-    "array_unique",
-    "array_flip",
-    "array_reverse",
-    "array_keys",
-    "array_values",
-    "array_map",
-    "array_filter",
-    "array_reduce",
-    "array_walk",
-    "array_fill",
-    "array_fill_keys",
-    "array_pad",
-    "sort",
-    "rsort",
-    "asort",
-    "arsort",
-    "ksort",
-    "krsort",
-    "usort",
-    "uasort",
-    "uksort",
-    "compact",
-    "extract",
-    "list",
-    "range",
-    // math
-    "abs",
-    "ceil",
-    "floor",
-    "round",
-    "max",
-    "min",
-    "pow",
-    "sqrt",
-    "log",
-    "exp",
-    "rand",
-    "mt_rand",
-    "random_int",
-    "fmod",
-    "intdiv",
-    "intval",
-    "floatval",
-    "is_nan",
-    "is_infinite",
-    "is_finite",
-    "pi",
-    "sin",
-    "cos",
-    "tan",
-    "asin",
-    "acos",
-    "atan",
-    "atan2",
-    // type / var
-    "isset",
-    "empty",
-    "unset",
-    "is_null",
-    "is_bool",
-    "is_int",
-    "is_integer",
-    "is_long",
-    "is_float",
-    "is_double",
-    "is_string",
-    "is_array",
-    "is_object",
-    "is_callable",
-    "is_numeric",
-    "is_a",
-    "instanceof",
-    "gettype",
-    "settype",
-    "intval",
-    "floatval",
-    "strval",
-    "boolval",
-    "var_dump",
-    "var_export",
-    "print_r",
-    "serialize",
-    "unserialize",
-    // file / io
-    "file_get_contents",
-    "file_put_contents",
-    "file_exists",
-    "is_file",
-    "is_dir",
-    "is_readable",
-    "is_writable",
-    "mkdir",
-    "rmdir",
-    "unlink",
-    "rename",
-    "copy",
-    "realpath",
-    "dirname",
-    "basename",
-    "pathinfo",
-    "glob",
-    "scandir",
-    "opendir",
-    "readdir",
-    "closedir",
-    "fopen",
-    "fclose",
-    "fread",
-    "fwrite",
-    "fgets",
-    "fputs",
-    "feof",
-    "fseek",
-    "ftell",
-    "rewind",
-    // date / time
-    "time",
-    "microtime",
-    "mktime",
-    "strtotime",
-    "date",
-    "date_create",
-    "date_format",
-    "date_diff",
-    "date_add",
-    "date_sub",
-    "checkdate",
-    // misc
-    "defined",
-    "define",
-    "constant",
-    "class_exists",
-    "interface_exists",
-    "function_exists",
-    "method_exists",
-    "property_exists",
-    "get_class",
-    "get_parent_class",
-    "is_subclass_of",
-    "header",
-    "headers_sent",
-    "setcookie",
-    "session_start",
-    "session_destroy",
-    "ob_start",
-    "ob_get_clean",
-    "ob_end_clean",
-    "json_encode",
-    "json_decode",
-    "call_user_func",
-    "call_user_func_array",
-    "array_walk_recursive",
-    "array_map",
-    "compact",
-    "extract",
-    "sleep",
-    "usleep",
-    "exit",
-    "die",
-];
-
-pub fn builtin_completions() -> Vec<CompletionItem> {
-    let mut seen = std::collections::HashSet::new();
-    PHP_BUILTINS
-        .iter()
-        .filter(|&&f| seen.insert(f))
-        .map(|f| callable_item(f, CompletionItemKind::FUNCTION, true))
-        .collect()
-}
-
-const PHP_SUPERGLOBALS: &[&str] = &[
-    "$_SERVER",
-    "$_GET",
-    "$_POST",
-    "$_FILES",
-    "$_COOKIE",
-    "$_SESSION",
-    "$_REQUEST",
-    "$_ENV",
-    "$GLOBALS",
-];
-
-pub fn superglobal_completions() -> Vec<CompletionItem> {
-    PHP_SUPERGLOBALS
-        .iter()
-        .map(|&name| CompletionItem {
-            label: name.to_string(),
-            kind: Some(CompletionItemKind::VARIABLE),
-            detail: Some("superglobal".to_string()),
-            ..Default::default()
-        })
-        .collect()
-}
-
-fn all_instance_members(
-    class_name: &str,
-    doc: &ParsedDoc,
-    other_docs: &[Arc<ParsedDoc>],
-) -> Vec<CompletionItem> {
-    let all: Vec<&ParsedDoc> = std::iter::once(doc)
-        .chain(other_docs.iter().map(|d| d.as_ref()))
-        .collect();
-    let mut items = Vec::new();
-    let mut seen_names: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut visited: std::collections::HashSet<String> = std::collections::HashSet::new();
-    // Queue: class names to process (inheritance chain + mixin chains).
-    let mut queue: Vec<String> = vec![class_name.to_string()];
-    while let Some(current) = queue.pop() {
-        if !visited.insert(current.clone()) {
-            continue;
-        }
-        let mut parent: Option<String> = None;
-        for d in &all {
-            let members = members_of_class(d, &current);
-            if parent.is_none() {
-                parent = members.parent.clone();
-            }
-            for (name, is_static) in members.methods {
-                if !is_static && seen_names.insert(name.clone()) {
-                    // Method params unknown here; use has_params=true so
-                    // snippet cursor lands inside parens.
-                    items.push(callable_item(&name, CompletionItemKind::METHOD, true));
-                }
-            }
-            for (name, is_static) in &members.properties {
-                if !is_static {
-                    let label = format!("${name}");
-                    if seen_names.insert(label.clone()) {
-                        let is_readonly = members.readonly_properties.contains(name);
-                        items.push(CompletionItem {
-                            label,
-                            kind: Some(CompletionItemKind::PROPERTY),
-                            detail: if is_readonly {
-                                Some("readonly".to_string())
-                            } else {
-                                None
-                            },
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            // Built-in enum properties: every enum case has `->name: string`
-            // and backed enums also have `->value`.
-            if is_enum(d, &current) {
-                if seen_names.insert("name".to_string()) {
-                    items.push(CompletionItem {
-                        label: "name".to_string(),
-                        kind: Some(CompletionItemKind::PROPERTY),
-                        detail: Some("string".to_string()),
-                        ..Default::default()
-                    });
-                }
-                if is_backed_enum(d, &current) && seen_names.insert("value".to_string()) {
-                    items.push(CompletionItem {
-                        label: "value".to_string(),
-                        kind: Some(CompletionItemKind::PROPERTY),
-                        detail: Some("string|int".to_string()),
-                        ..Default::default()
-                    });
-                }
-            }
-            // Collect @mixin classes for this class in this doc.
-            for mixin in mixin_classes_of(d, &current) {
-                queue.push(mixin);
-            }
-            // Queue trait names so their members are also included.
-            for trait_name in members.trait_uses {
-                queue.push(trait_name);
-            }
-        }
-        // Fall back to built-in stubs if the class wasn't found in any user doc
-        if let Some(stub) = builtin_class_members(&current) {
-            if parent.is_none() {
-                parent = stub.parent.clone();
-            }
-            for (name, is_static) in &stub.methods {
-                if !is_static && seen_names.insert(name.clone()) {
-                    items.push(callable_item(name, CompletionItemKind::METHOD, true));
-                }
-            }
-            for (name, is_static) in &stub.properties {
-                if !is_static {
-                    let label = format!("${name}");
-                    if seen_names.insert(label.clone()) {
-                        items.push(CompletionItem {
-                            label,
-                            kind: Some(CompletionItemKind::PROPERTY),
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            for name in &stub.constants {
-                if seen_names.insert(name.clone()) {
-                    items.push(CompletionItem {
-                        label: name.clone(),
-                        kind: Some(CompletionItemKind::CONSTANT),
-                        ..Default::default()
-                    });
-                }
-            }
-        }
-        if let Some(p) = parent {
-            queue.push(p);
-        }
-    }
-    items
-}
-
-fn all_static_members(
-    class_name: &str,
-    doc: &ParsedDoc,
-    other_docs: &[Arc<ParsedDoc>],
-) -> Vec<CompletionItem> {
-    let all: Vec<&ParsedDoc> = std::iter::once(doc)
-        .chain(other_docs.iter().map(|d| d.as_ref()))
-        .collect();
-    let mut items = Vec::new();
-    let mut seen_names: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut visited: std::collections::HashSet<String> = std::collections::HashSet::new();
-    let mut queue: Vec<String> = vec![class_name.to_string()];
-    while let Some(current) = queue.pop() {
-        if !visited.insert(current.clone()) {
-            continue;
-        }
-        let mut parent: Option<String> = None;
-        for d in &all {
-            let members = members_of_class(d, &current);
-            if parent.is_none() {
-                parent = members.parent.clone();
-            }
-            for (name, is_static) in members.methods {
-                if is_static && seen_names.insert(name.clone()) {
-                    items.push(callable_item(&name, CompletionItemKind::METHOD, true));
-                }
-            }
-            for (name, is_static) in members.properties {
-                if is_static {
-                    let label = format!("${name}");
-                    if seen_names.insert(label.clone()) {
-                        items.push(CompletionItem {
-                            label,
-                            kind: Some(CompletionItemKind::PROPERTY),
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            for name in members.constants {
-                if seen_names.insert(name.clone()) {
-                    items.push(CompletionItem {
-                        label: name,
-                        kind: Some(CompletionItemKind::CONSTANT),
-                        ..Default::default()
-                    });
-                }
-            }
-            // Queue trait names so their static members are also included.
-            for trait_name in members.trait_uses {
-                queue.push(trait_name);
-            }
-        }
-        // Fall back to built-in stubs for static members
-        if let Some(stub) = builtin_class_members(&current) {
-            if parent.is_none() {
-                parent = stub.parent.clone();
-            }
-            for (name, is_static) in &stub.methods {
-                if *is_static && seen_names.insert(name.clone()) {
-                    items.push(callable_item(name, CompletionItemKind::METHOD, true));
-                }
-            }
-            for (name, is_static) in &stub.properties {
-                if *is_static {
-                    let label = format!("${name}");
-                    if seen_names.insert(label.clone()) {
-                        items.push(CompletionItem {
-                            label,
-                            kind: Some(CompletionItemKind::PROPERTY),
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            for name in &stub.constants {
-                if seen_names.insert(name.clone()) {
-                    items.push(CompletionItem {
-                        label: name.clone(),
-                        kind: Some(CompletionItemKind::CONSTANT),
-                        ..Default::default()
-                    });
-                }
-            }
-        }
-        if let Some(p) = parent {
-            queue.push(p);
-        }
-    }
-    items
-}
-
-/// Resolve `ClassName::` or the aliases `self::`, `static::`, `parent::`.
-fn resolve_static_receiver(
-    source: &str,
-    doc: &ParsedDoc,
-    other_docs: &[Arc<ParsedDoc>],
-    position: Position,
-) -> Option<String> {
-    let line = source.lines().nth(position.line as usize)?;
-    let col = utf16_offset_to_byte(line, position.character as usize);
-    let before = &line[..col];
-    let before = before.strip_suffix("::").unwrap_or(before);
-    let name: String = before
-        .chars()
-        .rev()
-        .take_while(|&c| c.is_alphanumeric() || c == '_' || c == '\\')
-        .collect::<String>()
-        .chars()
-        .rev()
-        .collect();
-    match name.as_str() {
-        "" => None,
-        "self" | "static" => enclosing_class_at(source, doc, position),
-        "parent" => {
-            let enclosing = enclosing_class_at(source, doc, position)?;
-            // Look for the parent class in current doc then other docs
-            if let Some(p) = parent_class_name(doc, &enclosing) {
-                return Some(p);
-            }
-            for other in other_docs {
-                if let Some(p) = parent_class_name(other, &enclosing) {
-                    return Some(p);
-                }
-            }
-            None
-        }
-        _ => Some(name),
-    }
-}
-
 /// If the `(` trigger occurs inside an attribute like `#[ClassName(`, extract
 /// the attribute class name so we can offer its `__construct` parameter names.
 fn resolve_attribute_class(source: &str, position: Position) -> Option<String> {
@@ -994,181 +184,6 @@ fn resolve_call_params(
         }
     }
     params
-}
-
-/// Collect class/interface/trait/enum names with their FQN from an AST.
-/// Handles both braced (`namespace Foo { ... }`) and unbraced (`namespace Foo;`) forms.
-fn collect_classes_with_ns(
-    stmts: &[Stmt<'_, '_>],
-    ns_prefix: &str,
-    items: &mut Vec<(String, CompletionItemKind, String)>,
-) {
-    // `cur_ns` tracks the namespace context for unbraced `namespace Foo;` declarations,
-    // which apply to all subsequent statements at the same level.
-    let mut cur_ns = ns_prefix.to_string();
-
-    let fqn_for = |short: &str, ns: &str| -> String {
-        if ns.is_empty() {
-            short.to_string()
-        } else {
-            format!("{}\\{}", ns, short)
-        }
-    };
-
-    for stmt in stmts {
-        match &stmt.kind {
-            StmtKind::Class(c) => {
-                let short = c.name.unwrap_or("");
-                if !short.is_empty() {
-                    items.push((
-                        short.to_string(),
-                        CompletionItemKind::CLASS,
-                        fqn_for(short, &cur_ns),
-                    ));
-                }
-            }
-            StmtKind::Interface(i) => {
-                items.push((
-                    i.name.to_string(),
-                    CompletionItemKind::INTERFACE,
-                    fqn_for(i.name, &cur_ns),
-                ));
-            }
-            StmtKind::Trait(t) => {
-                items.push((
-                    t.name.to_string(),
-                    CompletionItemKind::CLASS,
-                    fqn_for(t.name, &cur_ns),
-                ));
-            }
-            StmtKind::Enum(e) => {
-                items.push((
-                    e.name.to_string(),
-                    CompletionItemKind::ENUM,
-                    fqn_for(e.name, &cur_ns),
-                ));
-            }
-            StmtKind::Namespace(ns) => {
-                let ns_name = ns
-                    .name
-                    .as_ref()
-                    .map(|n| n.to_string_repr().to_string())
-                    .unwrap_or_default();
-                match &ns.body {
-                    NamespaceBody::Braced(inner) => {
-                        collect_classes_with_ns(inner, &ns_name, items);
-                    }
-                    NamespaceBody::Simple => {
-                        // Unbraced namespace: applies to all subsequent statements.
-                        cur_ns = ns_name;
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// The line+col where a new `use` statement should be inserted in the current file.
-fn use_insert_position(source: &str) -> Position {
-    let mut last_use_line: Option<u32> = None;
-    let mut anchor_line: u32 = 0;
-    for (i, line) in source.lines().enumerate() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("<?") || trimmed.starts_with("namespace ") {
-            anchor_line = i as u32;
-        }
-        if trimmed.starts_with("use ") && !trimmed.starts_with("use function ") {
-            last_use_line = Some(i as u32);
-        }
-    }
-    Position {
-        line: last_use_line.unwrap_or(anchor_line) + 1,
-        character: 0,
-    }
-}
-
-/// The namespace declared at the top of the given statements, if any.
-fn current_file_namespace(stmts: &[Stmt<'_, '_>]) -> String {
-    for stmt in stmts {
-        if let StmtKind::Namespace(ns) = &stmt.kind {
-            return ns
-                .name
-                .as_ref()
-                .map(|n| n.to_string_repr().to_string())
-                .unwrap_or_default();
-        }
-    }
-    String::new()
-}
-
-const PHP_MAGIC_METHODS: &[(&str, &str)] = &[
-    (
-        "__construct",
-        "public function __construct($1)\n{\n    $2\n}",
-    ),
-    ("__destruct", "public function __destruct()\n{\n    $1\n}"),
-    (
-        "__get",
-        "public function __get(string $name): mixed\n{\n    $1\n}",
-    ),
-    (
-        "__set",
-        "public function __set(string $name, mixed $value): void\n{\n    $1\n}",
-    ),
-    (
-        "__isset",
-        "public function __isset(string $name): bool\n{\n    $1\n}",
-    ),
-    (
-        "__unset",
-        "public function __unset(string $name): void\n{\n    $1\n}",
-    ),
-    (
-        "__call",
-        "public function __call(string $name, array $arguments): mixed\n{\n    $1\n}",
-    ),
-    (
-        "__callStatic",
-        "public static function __callStatic(string $name, array $arguments): mixed\n{\n    $1\n}",
-    ),
-    (
-        "__toString",
-        "public function __toString(): string\n{\n    $1\n}",
-    ),
-    (
-        "__invoke",
-        "public function __invoke($1): mixed\n{\n    $2\n}",
-    ),
-    ("__clone", "public function __clone(): void\n{\n    $1\n}"),
-    ("__sleep", "public function __sleep(): array\n{\n    $1\n}"),
-    ("__wakeup", "public function __wakeup(): void\n{\n    $1\n}"),
-    (
-        "__serialize",
-        "public function __serialize(): array\n{\n    $1\n}",
-    ),
-    (
-        "__unserialize",
-        "public function __unserialize(array $data): void\n{\n    $1\n}",
-    ),
-    (
-        "__debugInfo",
-        "public function __debugInfo(): ?array\n{\n    $1\n}",
-    ),
-];
-
-fn magic_method_completions() -> Vec<CompletionItem> {
-    PHP_MAGIC_METHODS
-        .iter()
-        .map(|(name, snippet)| CompletionItem {
-            label: name.to_string(),
-            kind: Some(CompletionItemKind::METHOD),
-            insert_text: Some(snippet.to_string()),
-            insert_text_format: Some(InsertTextFormat::SNIPPET),
-            detail: Some("magic method".to_string()),
-            ..Default::default()
-        })
-        .collect()
 }
 
 /// Optional context for completion requests that enables richer results
@@ -1694,181 +709,6 @@ fn extract_match_subject(line: &str) -> Option<String> {
     } else {
         Some(var.to_string())
     }
-}
-
-/// Returns the prefix typed after `use ` on the current line, or None if not in a use statement.
-fn use_completion_prefix(source: &str, position: Position) -> Option<String> {
-    let line = source.lines().nth(position.line as usize)?;
-    let col = utf16_offset_to_byte(line, position.character as usize);
-    let before = line[..col].trim_start();
-    let prefix = before.strip_prefix("use ")?;
-    Some(prefix.trim_start_matches('\\').to_string())
-}
-
-/// Collect fully-qualified names from stmts that contain `prefix`.
-fn collect_fqns_with_prefix(
-    stmts: &[Stmt<'_, '_>],
-    ns: &str,
-    prefix: &str,
-    out: &mut Vec<CompletionItem>,
-) {
-    for stmt in stmts {
-        match &stmt.kind {
-            StmtKind::Class(c) => {
-                if let Some(name) = c.name {
-                    let fqn = if ns.is_empty() {
-                        name.to_string()
-                    } else {
-                        format!("{ns}\\{name}")
-                    };
-                    if fqn.to_lowercase().contains(&prefix.to_lowercase()) || prefix.is_empty() {
-                        out.push(CompletionItem {
-                            label: fqn.clone(),
-                            kind: Some(CompletionItemKind::CLASS),
-                            insert_text: Some(fqn),
-                            ..Default::default()
-                        });
-                    }
-                }
-            }
-            StmtKind::Interface(i) => {
-                let fqn = if ns.is_empty() {
-                    i.name.to_string()
-                } else {
-                    format!("{ns}\\{}", i.name)
-                };
-                if fqn.to_lowercase().contains(&prefix.to_lowercase()) || prefix.is_empty() {
-                    out.push(CompletionItem {
-                        label: fqn.clone(),
-                        kind: Some(CompletionItemKind::INTERFACE),
-                        insert_text: Some(fqn),
-                        ..Default::default()
-                    });
-                }
-            }
-            StmtKind::Namespace(ns_stmt) => {
-                let ns_name = ns_stmt
-                    .name
-                    .as_ref()
-                    .map(|n| {
-                        if ns.is_empty() {
-                            n.to_string_repr().to_string()
-                        } else {
-                            format!("{ns}\\{}", n.to_string_repr())
-                        }
-                    })
-                    .unwrap_or_else(|| ns.to_string());
-                if let NamespaceBody::Braced(inner) = &ns_stmt.body {
-                    collect_fqns_with_prefix(inner, &ns_name, prefix, out);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Extract the identifier characters typed immediately before the cursor.
-/// Includes `\` to support namespace-qualified prefixes like `App\Serv`.
-fn typed_prefix(source: Option<&str>, position: Option<Position>) -> Option<String> {
-    let src = source?;
-    let pos = position?;
-    let line = src.lines().nth(pos.line as usize)?;
-    let col = utf16_offset_to_byte(line, pos.character as usize);
-    let before = &line[..col];
-    let prefix: String = before
-        .chars()
-        .rev()
-        .take_while(|&c| c.is_alphanumeric() || c == '_' || c == '\\')
-        .collect::<String>()
-        .chars()
-        .rev()
-        .collect();
-    if prefix.is_empty() {
-        None
-    } else {
-        Some(prefix)
-    }
-}
-
-fn resolve_receiver_class(
-    source: &str,
-    doc: &ParsedDoc,
-    position: Position,
-    type_map: &TypeMap,
-) -> Option<String> {
-    let line = source.lines().nth(position.line as usize)?;
-    let col = utf16_offset_to_byte(line, position.character as usize);
-    let before = &line[..col];
-    // Try ?-> first (longer pattern) so `$s?->` doesn't get stripped to `$s?` by the `->` rule.
-    let before = before
-        .strip_suffix("?->")
-        .or_else(|| before.strip_suffix("->"))
-        .unwrap_or(before);
-
-    // Handle (new ClassName()) before ->
-    if let Some(class_name) = extract_new_class_before_arrow(before) {
-        return Some(class_name);
-    }
-
-    let var_name: String = before
-        .chars()
-        .rev()
-        .take_while(|&c| c.is_alphanumeric() || c == '_' || c == '$')
-        .collect::<String>()
-        .chars()
-        .rev()
-        .collect();
-    if var_name.is_empty() {
-        return None;
-    }
-    let var_name = if var_name.starts_with('$') {
-        var_name
-    } else {
-        format!("${var_name}")
-    };
-    if var_name == "$this" {
-        // Prefer the enclosing class (standard method context).
-        // Fall back to type_map for top-level bound closures where
-        // Closure::bind / bindTo / call injected a $this mapping.
-        return enclosing_class_at(source, doc, position)
-            .or_else(|| type_map.get("$this").map(|s| s.to_string()));
-    }
-    type_map.get(&var_name).map(|s| s.to_string())
-}
-
-/// Extract the class name from `(new ClassName(...))` or `new ClassName(...)` text
-/// appearing immediately before `->`.
-fn extract_new_class_before_arrow(text: &str) -> Option<String> {
-    let text = text.trim_end();
-    // Strip optional closing paren wrapping: `(new Foo())`
-    let inner = if let Some(without_last) = text.strip_suffix(')') {
-        // Find matching open paren — look for `(new` pattern
-        if let Some(pos) = without_last.rfind("(new ") {
-            &without_last[pos + 1..]
-        } else if let Some(pos) = without_last.rfind("(new\t") {
-            &without_last[pos + 1..]
-        } else {
-            text
-        }
-    } else {
-        text
-    };
-    // Now inner should start with `new ClassName(...)`
-    let inner = inner.trim();
-    if !inner.starts_with("new ") && !inner.starts_with("new\t") {
-        return None;
-    }
-    let after_new = inner[3..].trim_start();
-    // Extract class name (alphanumeric + _ + \)
-    let class: String = after_new
-        .chars()
-        .take_while(|&c| c.is_alphanumeric() || c == '_' || c == '\\')
-        .collect();
-    if class.is_empty() {
-        return None;
-    }
-    // Return short name
-    Some(class.rsplit('\\').next().unwrap_or(&class).to_string())
 }
 
 #[cfg(test)]
